% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubigeo_ccpp}
\alias{ubigeo_ccpp}
\title{UBIGEO for (Small) Population Centers (CCPP), INEI}
\format{
A data frame with 94922 rows and 9 variables:
\describe{
\item{\code{inei_ccpp}}{UBIGEO for the CCPP (INEI)}
\item{\code{inei_district}}{UBIGEO for the District (INEI)}
\item{\code{department}}{Name of the Department}
\item{\code{province}}{Name of the Province}
\item{\code{district}}{Name of the District}
\item{\code{ccpp}}{Name of the CCPP}
\item{\code{type}}{Type of the CCPP (Urban or Rural)}
\item{\code{latitude}}{Latitude}
\item{\code{longitude}}{Longitude}
}
}
\usage{
ubigeo_ccpp
}
\description{
This dataset contains information for small population centers (CCPP: Centros Poblados), including the INEI code, name, latitude and longitude.
}
\keyword{datasets}
